Move UnmanagedMemoryAccessor to shared CoreLib partition (#15137)
