+++ src/System.Private.CoreLib/src/System.Private.CoreLib.csproj
@@ -214,7 +214,6 @@
     <Compile Include="System\IO\InternalFile.Unix.cs" Condition="'$(TargetsUnix)'=='true'" />
     <Compile Include="System\IO\FileLoadException.CoreRT.cs" />
     <Compile Include="System\IO\Stream.cs" />
-    <Compile Include="System\IO\UnmanagedMemoryAccessor.cs" />
     <Compile Include="System\RuntimeMethodHandle.cs" />
     <Compile Include="System\Decimal.cs" />
     <Compile Include="System\Diagnostics\Contracts\Contracts.cs" />
